requirements.txt
Flask==2.3.2
pandas==2.2.3
statsmodel==0.14.0